+++ source/utils/utils.cpp
@@ -95,5 +95,5 @@ uint32_t getID(const char* txt) {
 
 float scale(float A, float A1, float A2, float Min, float Max) {
 	long double percentage = (A - A1) / (A1 - A2);
-	return (percentage) * (Min - Max) + Min;
+	return ((float)percentage) * (Min - Max) + Min;
 }
